update amount test and query order url

+++ src/main/java/pay2park/controller/payment/QueryPaymentUrlController.java
@@ -21,7 +21,7 @@ public class QueryPaymentUrlController {
 
     @GetMapping("/queryPaymentUrl")
     @ResponseBody
-    public ResponseObject createOrder(@RequestParam(value = "endUserId") int endUserID, @RequestParam(value = "ticketId") int ticketId ) throws IOException, URISyntaxException {
+    public ResponseObject createOrder(@RequestParam(value = "endUserId") int endUserID, @RequestParam(value = "ticketId") Long ticketId ) throws IOException, URISyntaxException {
         var data = queryPaymentUrlService.queryPaymentUrl(endUserID, ticketId);
